make deleteExpired() abstract
