+++ src/lib/keyStores/testMocks.ts
@@ -2,7 +2,6 @@
 
 import { derSerializePrivateKey } from '../crypto_wrappers/keys';
 import Certificate from '../crypto_wrappers/x509/Certificate';
-import { CertificateStore } from './CertificateStore';
 import { PrivateKeyData, PrivateKeyStore } from './privateKeyStore';
 import { PublicKeyStore, SessionPublicKeyData } from './publicKeyStore';
 
@@ -92,72 +91,5 @@ export class MockPublicKeyStore extends PublicKeyStore {
       throw new Error('Denied');
     }
     this.keys[peerPrivateAddress] = keyData;
-  }
-}
-
-export interface MockStoredCertificateData {
-  readonly expiryDate: Date;
-  readonly certificateSerialized: ArrayBuffer;
-}
-
-export class MockCertificateStore extends CertificateStore {
-  public dataByPrivateAddress: {
-    // tslint:disable-next-line:readonly-array
-    [privateAddress: string]: MockStoredCertificateData[];
-  } = {};
-
-  public expiredCertificatesDeleted: boolean = false;
-
-  public clear(): void {
-    this.dataByPrivateAddress = {};
-    this.expiredCertificatesDeleted = false;
-  }
-
-  public async forceSave(certificate: Certificate): Promise<void> {
-    await this.saveData(
-      await certificate.calculateSubjectPrivateAddress(),
-      certificate.serialize(),
-      certificate.expiryDate,
-    );
-  }
-
-  protected async deleteExpiredData(): Promise<void> {
-    this.expiredCertificatesDeleted = true;
-  }
-
-  protected async retrieveAllSerializations(
-    subjectPrivateAddress: string,
-  ): Promise<readonly ArrayBuffer[]> {
-    const certificateData = this.dataByPrivateAddress[subjectPrivateAddress];
-    if (!certificateData) {
-      return [];
-    }
-    return certificateData.map((d) => d.certificateSerialized);
-  }
-
-  protected async retrieveLatestSerialization(
-    subjectPrivateAddress: string,
-  ): Promise<ArrayBuffer | null> {
-    const certificateData = this.dataByPrivateAddress[subjectPrivateAddress] ?? [];
-    if (certificateData.length === 0) {
-      return null;
-    }
-    const dataSorted = certificateData.sort(
-      (a, b) => a.expiryDate.getDate() - b.expiryDate.getDate(),
-    );
-    return dataSorted[0].certificateSerialized;
-  }
-
-  protected async saveData(
-    subjectPrivateAddress: string,
-    subjectCertificateSerialized: ArrayBuffer,
-    subjectCertificateExpiryDate: Date,
-  ): Promise<void> {
-    const mockData: MockStoredCertificateData = {
-      certificateSerialized: subjectCertificateSerialized,
-      expiryDate: subjectCertificateExpiryDate,
-    };
-    const originalCertificateData = this.dataByPrivateAddress[subjectPrivateAddress] ?? [];
-    this.dataByPrivateAddress[subjectPrivateAddress] = [...originalCertificateData, mockData];
   }
 }
